Moved webcast selection dialog outside the cells; it now stands on its own
