+++ react/gameday2/components/VideoCell.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react'
 import RaisedButton from 'material-ui/RaisedButton'
 import WebcastEmbed from './WebcastEmbed'
 import VideoCellToolbarContainer from '../containers/VideoCellToolbarContainer'
-import WebcastSelectionOverlayDialogContainer from '../containers/WebcastSelectionOverlayDialogContainer'
+import WebcastSelectionDialogContainer from '../containers/WebcastSelectionDialogContainer'
 import SwapPositionOverlayDialogContainer from '../containers/SwapPositionOverlayDialogContainer'
 import { webcastPropType } from '../utils/webcastUtils'
 import { LAYOUT_STYLES } from '../constants/LayoutConstants'
@@ -73,7 +73,7 @@ export default class VideoCell extends React.Component {
             onRequestOpenWebcastSelectionDialog={() => this.onRequestOpenWebcastSelectionDialog()}
             onRequestOpenSwapPositionDialog={() => this.onRequestOpenSwapPositionDialog()}
           />
-          <WebcastSelectionOverlayDialogContainer
+        <WebcastSelectionDialogContainer
             open={this.state.webcastSelectionDialogOpen}
             webcast={this.props.webcast}
             onRequestClose={() => this.onRequestCloseWebcastSelectionDialog()}
@@ -116,7 +116,7 @@ export default class VideoCell extends React.Component {
             onTouchTap={() => this.onRequestOpenWebcastSelectionDialog()}
           />
         </div>
-        <WebcastSelectionOverlayDialogContainer
+        <WebcastSelectionDialogContainer
           open={this.state.webcastSelectionDialogOpen}
           webcast={this.props.webcast}
           onRequestClose={() => this.onRequestCloseWebcastSelectionDialog()}
