+++ react/gameday2/components/WebcastSelectionDialog.js
@@ -1,4 +1,5 @@
 import React, { PropTypes } from 'react'
+import Dialog from 'material-ui/Dialog'
 import Paper from 'material-ui/Paper'
 import { green500, indigo500 } from 'material-ui/styles/colors'
 import Divider from 'material-ui/Divider'
@@ -10,10 +11,10 @@ import ActionHelp from 'material-ui/svg-icons/action/help'
 import VideoCam from 'material-ui/svg-icons/av/videocam'
 import VideoCamOff from 'material-ui/svg-icons/av/videocam-off'
 import EventListener from 'react-event-listener'
-import WebcastSelectionOverlayDialogItem from './WebcastSelectionOverlayDialogItem'
+import WebcastSelectionDialogItem from './WebcastSelectionDialogItem'
 import { webcastPropType } from '../utils/webcastUtils'
 
-export default class VideoCellOverlayDialog extends React.Component {
+export default class WebcastSelectionDialog extends React.Component {
 
   static propTypes = {
     open: PropTypes.bool.isRequired,
@@ -27,18 +28,6 @@ export default class VideoCellOverlayDialog extends React.Component {
 
   constructor(props) {
     super(props)
-
-    this.layout = {
-      dialogMargin: 20,
-    }
-  }
-
-  componentDidMount() {
-    this.updateSizing()
-  }
-
-  componentDidUpdate() {
-    this.updateSizing()
   }
 
   onRequestClose() {
@@ -47,77 +36,7 @@ export default class VideoCellOverlayDialog extends React.Component {
     }
   }
 
-  updateSizing() {
-    if (this.props.open) {
-      const component = this.component
-      const dialogListContainer = this.dialogListContainer
-      const dialogList = this.dialogList
-
-      let dialogHeight = 0
-      dialogHeight += dialogListContainer.previousSibling.offsetHeight
-      dialogHeight += dialogListContainer.nextSibling.offsetHeight
-      dialogHeight += dialogList.offsetHeight
-
-      const maxHeight = component.offsetHeight - (2 * this.layout.dialogMargin)
-      if (dialogHeight > maxHeight) {
-        dialogListContainer.style.overflowY = 'auto'
-        let listContainerHeight = maxHeight
-        listContainerHeight -= dialogListContainer.previousSibling.offsetHeight
-        listContainerHeight -= dialogListContainer.nextSibling.offsetHeight
-        dialogListContainer.style.height = `${listContainerHeight}px`
-      } else {
-        dialogListContainer.style.height = null
-      }
-    }
-  }
-
   render() {
-    const wrapperStyle = {
-      position: 'absolute',
-      top: 0,
-      bottom: 0,
-      left: 0,
-      right: 0,
-      backgroundColor: 'rgba(0, 0, 0, 0.541176)',
-    }
-
-    const paperStyle = {
-      margin: this.layout.dialogMargin,
-      width: 'calc(100% - 40px)',
-      maxWidth: '400px',
-      marginLeft: 'auto',
-      marginRight: 'auto',
-      position: 'relative',
-    }
-
-    const titleStyle = {
-      padding: '16px',
-      fontSize: '22px',
-      margin: 0,
-      fontWeight: 400,
-      borderBottom: '1px solid rgb(224, 224, 224)',
-    }
-
-    const buttonContainerStyle = {
-      padding: '8px',
-      width: '100%',
-      textAlign: 'right',
-      borderTop: '1px solid rgb(224, 224, 224)',
-    }
-
-    const listContainerStyle = {
-      overflowY: 'auto',
-    }
-
-    const buttonContainerElement = (
-      <div style={buttonContainerStyle}>
-        <FlatButton
-          label="Cancel"
-          onTouchTap={() => this.onRequestClose()}
-        />
-      </div>
-    )
-
     const subheaderStyle = {
       color: indigo500,
     }
@@ -146,7 +65,7 @@ export default class VideoCellOverlayDialog extends React.Component {
 
       if (this.props.specialWebcastIds.has(webcast.id)) {
         const item = (
-          <WebcastSelectionOverlayDialogItem
+          <WebcastSelectionDialogItem
             key={webcast.id}
             webcast={webcast}
             webcastSelected={this.props.onWebcastSelected}
@@ -160,7 +79,7 @@ export default class VideoCellOverlayDialog extends React.Component {
         }
       } else if (webcast.id.startsWith('bluezone')) {
         bluezoneWebcastItems.push(
-          <WebcastSelectionOverlayDialogItem
+          <WebcastSelectionDialogItem
             key={webcast.id}
             webcast={webcast}
             webcastSelected={this.props.onWebcastSelected}
@@ -170,7 +89,7 @@ export default class VideoCellOverlayDialog extends React.Component {
         )
       } else {
         const item = (
-          <WebcastSelectionOverlayDialogItem
+          <WebcastSelectionDialogItem
             key={webcast.id}
             webcast={webcast}
             webcastSelected={this.props.onWebcastSelected}
@@ -259,42 +178,37 @@ export default class VideoCellOverlayDialog extends React.Component {
       )
     }
 
-    if (this.props.open) {
-      // This "div" soup is needed because React is deprecating the ability to
-      // access the DOM nodes of components, so we need to wrap them in divs in
-      // order to measure and size them. See https://github.com/yannickcr/eslint-plugin-react/issues/678
-      return (
-        <div
-          style={wrapperStyle}
-          onTouchTap={() => this.onRequestClose()}
-          ref={(e) => { this.component = e }}
-        >
-          <EventListener
-            target="window"
-            onResize={() => this.updateSizing()}
-          />
-          <Paper
-            style={paperStyle}
-            zDepth={5}
-            onTouchTap={(e) => e.stopPropagation()}
-          >
-            <h3 style={titleStyle}>Select a webcast</h3>
-            <div
-              style={listContainerStyle}
-              ref={(e) => { this.dialogListContainer = e }}
-            >
-              <div ref={(e) => { this.dialogList = e }}>
-                <List>
-                  {allWebcastItems}
-                </List>
-              </div>
-            </div>
-            {buttonContainerElement}
-          </Paper>
-        </div>
-      )
+    const actions = [
+      <FlatButton
+        label="Cancel"
+        onTouchTap={() => this.onRequestClose()}
+        primary={true}
+      />,
+    ]
+
+    const titleStyle = {
+      padding: 16,
     }
 
-    return null
+    const bodyStyle = {
+      padding: 0,
+    }
+
+    return (
+      <Dialog
+        title="Select a webcast"
+        actions={actions}
+        modal={false}
+        titleStyle={titleStyle}
+        bodyStyle={bodyStyle}
+        open={this.props.open}
+        onRequestClose={() => this.onRequestClose()}
+        autoScrollBodyContent={true}
+      >
+        <List>
+          {allWebcastItems}
+        </List>
+      </Dialog>
+    )
   }
 }
