+++ react/gameday2/components/WebcastSelectionDialogItem.js
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react'
 import { ListItem } from 'material-ui/List'
 
-export default class WebcastSelectionOverlayDialogItem extends React.Component {
+export default class WebcastSelectionDialogItem extends React.Component {
   static propTypes = {
     webcast: PropTypes.object.isRequired,
     webcastSelected: PropTypes.func.isRequired,
