--- react/gameday2/containers/WebcastSelectionOverlayDialogContainer.js
@@ -1,20 +1,0 @@
-import { connect } from 'react-redux'
-import WebcastSelectionOverlayDialog from '../components/WebcastSelectionOverlayDialog'
-import { addWebcastAtPosition, swapWebcasts, removeWebcast } from '../actions'
-import { getWebcastIdsInDisplayOrder } from '../selectors'
-
-const mapStateToProps = (state) => ({
-  webcasts: getWebcastIdsInDisplayOrder(state),
-  webcastsById: state.webcastsById,
-  specialWebcastIds: state.specialWebcastIds,
-  displayedWebcasts: state.videoGrid.displayed,
-  layoutId: state.videoGrid.layoutId,
-})
-
-const mapDispatchToProps = (dispatch) => ({
-  removeWebcast: (id) => dispatch(removeWebcast(id)),
-  addWebcastAtPosition: (webcastId, position) => dispatch(addWebcastAtPosition(webcastId, position)),
-  swapWebcasts: (firstPosition, secondPosition) => dispatch(swapWebcasts(firstPosition, secondPosition)),
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(WebcastSelectionOverlayDialog)
